+++ intellij-plugin/src/main/kotlin/com/ymkang/gitcryptexplorer/ui/GitCryptIcons.kt
@@ -1,0 +1,9 @@
+package com.ymkang.gitcryptexplorer.ui
+
+import com.intellij.openapi.util.IconLoader
+import javax.swing.Icon
+
+object GitCryptIcons {
+    @JvmField
+    val Explorer: Icon = IconLoader.getIcon("/icons/git-crypt-explorer.svg", GitCryptIcons::class.java)
+}